+++ changelog.md
@@ -1,3 +1,12 @@
+1.0.3.0 2021-07-31
+==================
+- add parseOpts', which hopefully has better `hsubparser` handling
+
+1.0.2.0 2020-09-17
+==================
+- better help handling
+- add facilities for working with Options.Applicative.Help.Pretty.Doc
+
 1.0.1.0 2020-01-06
 ==================
 - argS -> argT, add optT